+++ app/src/main/java/com/madrid/movio/app.kt
@@ -6,7 +6,7 @@ import com.madrid.data.repositories.MovieDetailsRepositoryImpl
 import com.madrid.data.repositories.RecommendedRepositoryImp
 import com.madrid.data.repositories.SearchRepositoryImpl
 import com.madrid.data.repositories.local.LocalDataSource
-import com.madrid.data.repositories.local.LocalDataSourceImpl
+import com.madrid.data.dataSource.local.LocalDataSourceImpl
 import com.madrid.detectImageContent.GetImageBitmap
 import com.madrid.detectImageContent.SensitiveContentDetection
 import com.madrid.domain.repository.MovieDetailsRepository
